feat: adding actors validations

+++ src/actor/actor.service.ts
@@ -9,11 +9,21 @@ export class ActorService {
   }
 
   public list(): IActor[] {
-    return this.actorRepository.findAll();
+    const actors = this.actorRepository.findAll();
+
+    if(!actors || actors.length === 0) {
+      throw new Error("No actors found in database.");
+    }
+    return actors
   }
 
   public getById(id: number): IActor | undefined {
-    return this.actorRepository.findById(id);
+    const actor = this.actorRepository.findById(id);
+
+    if (!actor) {
+      throw new Error(`Actor with ${id} not found in database.`);
+    }
+    return actor
   }
 
   public create(actor: Omit<IActor, "id">): IActor {
@@ -21,10 +31,20 @@ export class ActorService {
   }
 
   public update(id: number, actorData: Partial<Omit<IActor, "id">>): IActor | undefined {
+    const actor = this.actorRepository.findById(id);
+
+    if (!actor) {
+      throw new Error(`Actor with ${id} not found in database. Unable to update.`);
+    }
     return this.actorRepository.update(id, actorData);
   }
 
   public delete(id: number): void {
+    const actor = this.actorRepository.findById(id);
+
+    if (!actor) {
+      throw new Error(`Actor with ${id} not found in database. Unable to delete.`);
+    }
     this.actorRepository.delete(id);
   }
 }
